Employee main conflict resolve

+++ backend/src/main/java/net/primepro/primepro/dto/ChangeStatusDto.java
@@ -8,6 +8,6 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @Data
 public class ChangeStatusDto {
-    private Long taskId;
+    private Integer taskId;
     private String status;
 }
